Throw an Error when Mirage route is not defined
While TDDing, test failure feedback is very important. We prefer loud
failures of quiet ones.

Unfortunately, `console.log` won't cause a test failure while running
`ember test --serve`. If a Mirage route is undefined, tests tend to
hang.

Upon further investigation in a UI-backed browser like Chrome or
Firefox, the `console.log` message is visible.

This feedback is still useful, but we'd prefer to see it in the console,
without having to resort to a visual browser.