Don't mutate model classes
Model classes with associations can only be correctly registered with a
Schema once, because the registration process mutates them. This change
extends them instead so that we don't mutate the originals.

I needed to switch to `for ... in` instead of `Object.keys` because now
we definitely need to walk the prototype chain to find
associations. This was probably a separate latent bug, since users can
put associations onto ancestor classes.