Improve ID Auto-Incrementation
This introduces a helper class to manage the value of a collection's ID
when inserting records.

ID values are also tracked to ensure we avoid collisions. An error will
be thrown when attempting to insert a record whose ID has been used. An
error will also be thrown if you attempt to update the ID of an existing
record.

Resolves #415