Ensure links work with multiple relationships
The `link` serializer function was incorrectly firing every iteration of
the relationships, which caused the last link to be assigned for
everything. We now get the link data once before relationships are
processed, and refer to it when iterating.

Refactored the test to process multiple relationships to ensure this
doesn't happen again.