Add README for main and convergence package.
The docs are extraordinarily bare right now, but now that this is a
pinned repository they need at least a guide of what we're working on,
and how that fits into the overall project.

This adds two pieces of documentation, a README for the main repo, and
a README for the @bigtest/convergence package. The main repo README
contains an outline of the philosophy of big testing, and inserts at
appropriate points, the readmes into the specific packages that
address that philosophy. For example, the link to
`@bigtest/convergence` falls under the "Real Asynchrony" section.

The README for the convergence package sticks to the high level of
"why" convergence is necessary, without actually mentioning the actual
API yet since that is still in the works. The idea is that the API
docs can be added after the justification when they are ready.