Tighten up collection property
By not returning a property descriptor, the method will only be made
available on page-objects and not their custom interactions. This is
so interaction methods will always return new interactions.

Added an error when the element does not exist at the specified
index. When a page-object is given `undefined` as the root selector,
the default `document.body` element is used as the page-object root
element. This causes all subsequent collection methods to search the
entire document.