Refactor convergent assertions to account for latency
A latency method that acts similarly to mocha's timeout method will be
added to every test and hook created within a describe block using
this package's exported describe function. When convergences are
returned from hooks, or when convergent assertions are run, they use
the test's current timeout to converge. To allow throwing errors after
the timeout, and allow `always` convergences to utilize this timeout,
the test/hook's timeout is increased by the latency amount. By
default, this is 100ms.