Fix each hook timeout error
When an each hook is run more than once, it retains the old
timeout. When our helper sets the timeout to 0, this causes the next
run to give the convergence a timeout of 0.

By setting the timeout back to the original timeout before the
convergence finishes, we can ensure that the next run of the hook
still has the correct timeout.