🐛Remove event handler in teardown hook
In order versions of effection, the teardown hook was added by
returning a function from the controller. The newest version however,
requires explicitly invoking the `ensure` function in order to
register the hook. We didn't update to the new method, and so our
event handlers were never getting un-registered.

This wasn't effecting us because the `resume`() of a halted context is
a no-op, but it was leaking the context itself which is not good.