Use separate mailboxes for each component's delegates
This way it is much clearer where messages are received. The mailbox is no longer shared between the orchestrator and other parts of the system. The orchestrator's own delegate cannot receive messages from its subcomponents. These changes clarify the following definition of a delegate in this system:

> A delegate is a mailbox where the caller (the outer process) receives messages, and the callee (the component which receives the delegate as an option) sends messages.

We will also want the reverse, pretty soon, where the caller (or someone else) is the sender, and the callee is the recipient. We will want this for the connection server, to initiate test runs in the connected agents, for example. I propose we call this concept "inbox" to separate it from a "delegate".