Add @bigtest/agent
Motivation
----------

There are a couple of problems with our current setup of having the
agent living inside the bigtest/server.

First and foremost, we're building the agent application code every
single time and that causes a lot of unwanted startup time. We're
willing to deal with this in development, but this is really going to
add significant time to work with bigtest applications in real life,
and also consume lots of unecessary computational resources in CI.

Approach
--------

This extracts the agent server app into a pre-built web application
that uses parcel and microbundle to compile the HTML and typescript
respectively.

In addition, an enhancement has been made to have the agent server
bind to optionally bind to a different port every time.