Fingerprint manifest so they get unique URLs
The fingerprinted filename is stored in the state.

One minor change in this PR is that the orchestrator no longer creates the atom, but must be supplied with one. The reason for this is that we need to be able to extract data from the atom in our tests, and there is no other way of getting hold of it.

Depends on #101