Upgrade agent to use effection 0.5.1
Prior to 0.5.1, effection will not work with `regeneratorRuntime`
since it used native generator function prototype to test if an object
was a generator. Since the agent needs to support any old browser, it
uses `regeneratorRuntime` instead of native generators.

This upgrades the agent to the first version of effection that uses
`regeneratorRuntime`