Use atomic slices to improve type-checking of lens operations
Handy as they are, using TypeScript with Ramda lenses can be annoying
at best, and perilous at worst. There are two basic problems:

1. lens operations are cumbersome in that you have to store a
reference to not only the child and the root value, but also the lens.
2. the lenses erase type information. Ideally they shouldn't since
they don't in other statically typed languages, but they don't type
easily since TS does not support higher kinded types.

This solves both of those problem by capturing a lens into a `Slice`
object. That way, there's no need to construct an extra referenc, and
all of the operations on that slice are implicity happening on the
focus of that lens. It solves the typing problem by the fact that
`Slice`s are parameterized with the type of the lens's focus. While it
isn't 100% correct, it does side-step the issue of typing lenses
entirely by essentially declaring at the point you take the slice
"this slice has this type" and while that is not strictly type-safe,
if you do get that right, then every subsequent usage of the slice
thereafter _is_.