Add `@bigtest/parcel` package to control parcel processes
Motivation
----------

We can't ship bigtest with a runtime that is dependent on ts-node. Not
only is it way too gigantic, it takes too long to start up if first we
have to compile all the TypeScript on every start up.

Approach
--------

This takes a step towards ts-node independence by extracting the
parcel process into its own package that can be pre-compiled and
imported by the `@bigtest/server`

The ParcelProcess is an object that, when allocated with the
`create()` operation, will make sure that as soon as the current
operation passes out of scope, the entire parcel worker farm is torn
down, no matter what stage of readiness it is in.

Also, the parcel process implements a `receive()` method so that you
can await updates from it.