Make server distributable via NPM
Motivation
----------

Up to this point we have not been able to precompile the server code
to JavaScript because it has had several runtime dependencies on
ts-node; the primary one of these being the parcel process that
bundles the tests. Because we can't precompile the server, it means
that we can't distribute it to NPM.

However, now because we have `@bigtest/parcel` as a separate package,
we can compile the server and upload it to NPM so that it can be used
as a development dependency of any project!

Approach
--------

1. Add a prepack script to compile our typescript similar to the way
we compile our other projects
2. flip the switch to start publishing to NPM