Use built assets as package contents
Even using the prepack script, the server package does not include the
compiled assets because they are not included in the files property of
the package.json That means that the actual code is not shipped to NPM
and you end up with `Error: Cannot find module './orchestrator'` when
trying to start bigtest server.

This explicitly distributes all of the files in `dist/` along with
their typescript declarations with NPM package.