Use external iteration for client subscription
Currently a client subscription uses a form of internal iteration like this:

```
subscribe("{ foo }", function*(data) {
  // do something
});
```

This changes the subscription to use external iteration instead, like this:

```
let subscription = subscribe("{ foo }");

while(true) {
  let data = yield subscription.receive();
  // do something
}
```

The major advantage to this is that we can at some point decide we are done, whereas we have no good way of doing this with internal iteration. With internal iteration we have to surreptitiously obtain a handle to the current ExecutionContext and halt it, which is pretty gross.

We will need this for the CLI, since a CLI command will fire off a subscription, but must be able to decide that it is now done at some point.