Create @bigtest/webdriver
Motivation
----------

BigTest needs to be able to start and stop browsers automatically in
order to create the browser agents that will be part of a test
run. So, if we want to run our test suite on Firefox _and_ Chrome
within the single test run, we'll need to be able to start those
browsers, and then direct them at the agent server.

Approach
--------

All the major browsers support the [webdriver protocol][1], both to
startup and to control a browser over a series of HTTP
requests. Rather than try and implement our own control mechanisms,
this uses webdriver to launch a managed browser and have it navigate
to a url.

This arrangement is ideal because most of the actual direction of the
test happens within the browser itself as part of the agent/harness
combination found in the `@bigtest/agent` package. In fact, the only
operation that we need currently beyond starting and stopping is the
`WebDriver#navigateTo()` which loads a url inside of the browser under
control.

In this change, we introduce the subset of webdriver functionality
that we need in order to get a bigtest running. This can be found in
the `WebDriver` interface. Then, we provide a `Local()` operation to
create a `WebDriver` that uses a locally installed Chrome or a Firefox
to actually run the session.

For example:

```typescript
import { Operation } from 'effection';
import { WebDriver, Local } from '@bigtest/webdriver';

export function* connect(url: string): Operation<WebDriver> {
  // get the instance of the driver
  let driver: WebDriver = yield Local('chromedriver', { headless: true });

  // use it to got to a web page
  yield driver.navigateTo('https://frontside.com');

  // pass the driver resource to the caller
  return driver;
}
```

Questions
---------

**Q:** Does this use Selenium?

**A:** No. This implementation does not use [Selenium][5]. It uses the
  webdriver protocol *directly* over HTTP.  As a result, it is much
  faster, and more tightly integrated with the [`effection`][4]
  paradigm which makes it more robust to extremely rapid setup and
  teardown.

***Q:** What about remote browsers launched from a cloud service?

**A:** While the scope of this change only includes launching local
browser processes and not connecting to remote webdriver servers such
as [BrowserStack][2] or [Sauce Labs][3], the implementation should be
relatively straightforward, since even communication with the local
processes happens over HTTP using webdriver REST calls.

**Q:** What about browsers and devices that don't support webdriver?
**A:** Between all the various cloud services it is actually quite
  difficult to find a browser that cannot be launched via
  webdriver. For those that don't, our useage of the protocol is very
  slim. In fact, it is limited to three REST calls: `POST /session`,
  `DELETE /session/${id}`, and `POST /session/${id}/url`. In the event
  that a device does not support webdriver, emulating those three
  calls should be within the realm of possibility. Finally, remember
  that BigTest does not require that it be the one to launch a browser
  in order for that browser to be an agent, only that the browser
  connect to the agent server url. It will always support just opening
  a browser and entering in the agent connect url.

[1]: https://w3c.github.io/webdriver/
[2]: https://www.browserstack.com
[3]: https://saucelabs.com
[4]: https://github.com/thefrontside/effection
[5]: https://www.selenium.dev