Automatically launch agents from the orchestrator/cli
We want to be able to weave the starting and stopping of browsers into
test flow seamlessly, whether in CI, or during development.

This adds the capability to manage a set of browsers to the
orchestrator, and then specify which managed browsers to launch when
the orchestrator spins up.

The tack we take is to create a set named of "drivers" associated with
the project that specify how to launch a runtime that will contain an
agent so that tests can be run. At this moment, "driver" is really a
synonym for "webdriver local", but the vocabulary was left intentionally
more abstract in order to accomodate other kinds of drivers such as
remote webdrivers as well as iOS, Android, and Node drivers.

In the `.bigtest.json` config, this adds the "drivers" field and a
"launch" field. Each entry in the "drivers" field specifies a driver
configuration that can be used by the "launch" field (and also by the
`--launch` command line option). The idea is that the "driver" is just
an opaque blop of configuration that can be passed to the underlying
driver when it is launched. For example, you might have two drivers
for chrome, one that uses the default window size, and another that
launches with a small window in order to test the small breakpoint in
a responsive layout:

```json
"drivers": {
  "chrome": {
    "browserName": "chrome",
    "headless": true,
  },
  "chrome-small": {
    "browserName": "chrome",
    "headless": true,
    "width": 600
  },
  "launch": {
    "chrome": {}
  }
}
```

Now, when starting the server, the orchestrator will see the
"chrome" option in the project config's "launch" field, and
automatically launch that driver with the configuration in the driver
field merged with whatever is in the launch field. While
"chrome-small" is _available_ to be launched, it will not be launced
by default in the absence of a `--launch` parameter. However, if you
_do_ want to launch it, here's how you would:

```
$ bigtest server --launch chrome-small
```

In order to facilitate having a single opaque configuration parameter,
the `@bigtest/webdriver` package was refactored to take a single
parameter.

TODOs and Open Questions:
-------------------------

- [ ] Not super happy with the additions to Project config. The
  problem is, that from the command line you want to be able override
  configuration like setting "headless" to `false`, so the format of
  the "launch" field has to have the same format as the "driver"
  field, but that's somewhat confusing since the purpose of the
  "launch" field isn't to provide configuration, just specify _which_
  browsers to launch. Feels more like it should be an array like
  `"launch": ["chrome"']``. I only did it this way so that
  `createServer` could take a `ProjectOptions` object and not have to
  diverge.
- [ ] The argument parsing for the `--launch` flag feels really quite
  flimsy. It parses everything after the `@` sign as [`JSON5`][1] with
  an implicit bracket, which feels hackish, but more importantly it
  lacks the things you expect from a robust CLI like:
    - at a minumum validate of options, so `--launch chrome@blah:true`
      is an error as would be `--launch chrome@headless:5`
    - completition: `--launch X` should complete X from the list of
      available drivers. `--launch chrome@` should should completation
      of the chrome drivers attributes, etc..
- [ ] Don't want to make the config confusing so that it's constantly
  a source of frustration. Still, drivers will have complex
  configuration, especially for remote cloud-based drivers that will
  need to specify driver urls and auth tokens. Maybe making
  configuration sparse?
- [ ] There is a potential race condition lurking in the browser
  launching resource. According to the webdriver spec, the navigation
  command will return once the `load` event has been dispatched on the
  document. This makes sense, but we really don't want to consider the
  launch complete until we receive the corresponding `connect` event
  from the agent as it calls home to activate itself. If we try to run
  the test before this connect event happens, or worse, if the agent
  code somehow fails after the `load` event happens, we'll think the
  agent should be connected when in fact it is not.

[1]: https://json5.org